+++ src/route/(components)/input.tsx
@@ -1,10 +1,12 @@
-import { InputHTMLAttributes } from 'react';
+import ErrorIcon from '@Assets/icons/Error';
+import { InputHTMLAttributes, useState } from 'react';
+import { FieldError } from 'react-hook-form';
 
 interface InputProps extends InputHTMLAttributes<HTMLInputElement> {
   name: string;
-  children: JSX.Element;
+  children?: JSX.Element;
   callback?: (e: React.ChangeEvent<HTMLInputElement>) => void;
-  error: string;
+  error?: FieldError | string | undefined;
   register?: any;
 }
 
@@ -16,26 +18,50 @@ export default function Input({
   register,
   ...rest
 }: InputProps) {
+  const [hover, setHover] = useState(false);
+
+  const handleMouseEnter = () => setHover(true);
+  const handleMouseLeave = () => setHover(false);
+
   return (
-    <label className="relative flex flex-col m-auto h-full">
-      <span className="absolute inset-y-0 left-0 flex items-center pl-2 grow-0">
-        {children}
-      </span>
-      <input
-        className={
-          'placeholder:italic block bg-slate-700 w-full rounded-md py-2 pl-10 pr-3 shadow-sm focus:outline-none lining-nums tabular-nums grow-1 h-fit ' +
-          rest.className
-        }
-        placeholder={`${rest.placeholder || name}...`}
-        type="text"
-        onChange={callback}
-        {...(register
-          ? register(name, {
-              valueAsNumber: rest.type === 'number',
-            })
-          : undefined)}
-      />
-      <span className="text-red-500">{error}</span>
-    </label>
+    <div
+      className={
+        'm-auto bg-slate-700 rounded-md shadow-sm h-fit py-1 px-2 lining-nums tabular-nums w-[70%] ' +
+        (error ? ' ring-offset-4 ring-offset-slate-800 ring-2 ring-red-500' : '')
+      }
+    >
+      <div className="flex">
+        {children ? (
+          <span className="flex justify-center items-start">{children}</span>
+        ) : null}
+        <div className="grow py-2 px-1">
+          <input
+            className={
+              'placeholder:italic bg-slate-700 rounded-md focus:outline-none px-1 w-full ' +
+              rest.className +
+              (error && hover ? ' w-0' : '')
+            }
+            placeholder={`${rest.placeholder || name}...`}
+            type="text"
+            onChange={callback}
+            {...(register
+              ? register(name, {
+                  valueAsNumber: rest.type === 'number',
+                })
+              : undefined)}
+          />
+          {error && hover && (
+            <span className="text-red-500">{typeof error !== 'string' ? `${error.message}: ${error.type}`:error}</span>
+          )}
+        </div>
+        <span className="flex justify-center items-start">
+          <ErrorIcon
+            className={"h-6 w-6 m-auto " + (error ? 'fill-red-500' : ' fill-transparent')}
+            onMouseEnter={handleMouseEnter}
+            onMouseLeave={handleMouseLeave}
+          />
+        </span>
+      </div>
+    </div>
   );
 }